+++ src/records/Webhook.php
@@ -11,6 +11,7 @@ use craft\db\ActiveRecord;
  * @property int|null $groupId
  * @property string $name
  * @property string $class
+ * @property string $filters
  * @property string $event
  * @property string $url
  * @property string|null $userAttributes